refactored jcache sample create-cache to configure-cache-types
