+++ jcache/configure-cache-types/src/test/resources/beans.xml
