+++ jcache/configure-cache-types/src/test/resources/jboss-deployment-structure.xml
