--- jcache/create-cache/src/main/java/org/javaee8/jcache/core/KeyValueService.java
@@ -1,42 +1,0 @@
-package org.javaee8.jcache.core;
-
-import javax.annotation.PostConstruct;
-import javax.annotation.PreDestroy;
-import javax.cache.Cache;
-import javax.cache.CacheManager;
-import javax.cache.Caching;
-import javax.cache.configuration.MutableConfiguration;
-import javax.cache.spi.CachingProvider;
-
-
-/**
- * @author Radim Hanus
- */
-public class KeyValueService<K,V> {
-    private Cache<K, V> cache;
-
-    @PostConstruct
-    void init() {
-        CachingProvider cachingProvider = Caching.getCachingProvider();
-        CacheManager cacheManager = cachingProvider.getCacheManager();
-        MutableConfiguration<K, V> config = new MutableConfiguration<>();
-        cache = cacheManager.createCache("cache.default", config);
-    }
-
-    @PreDestroy
-    void destroy() {
-        cache.close();
-    }
-
-    public void put(K key, V value) {
-        cache.put(key, value);
-    }
-
-    public V get(K key) {
-        return cache.get(key);
-    }
-
-    public void remove(K key) {
-        cache.remove(key);
-    }
-}
